+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.0 IN PROGRESS
 
 * Migrate from react-hot-loader to react-refresh. Refs STRWEB-27.
+* `autoprefixer` and `postcss` versions are now compatible. Refs STRWEB-46.
 
 ## [3.0.3](https://github.com/folio-org/stripes-webpack/tree/v3.0.3) (2022-02-10)
 [Full Changelog](https://github.com/folio-org/stripes-webpack/compare/v3.0.2...v3.0.3)