+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for stripes-webpack
 
+## 4.1.0 IN PROGRESS
+
 ## [4.0.0](https://github.com/folio-org/stripes-webpack/tree/v4.0.0) (2022-06-14)
 [Full Changelog](https://github.com/folio-org/stripes-webpack/compare/v3.0.3...v4.0.0)